+++ data-raw/exampleData.R
@@ -1,9 +1,8 @@
 ## Create TDHeat05.
 
 # Read raw data
-SB_Yield <- read.csv(system.file("extdata", "SB_yield.csv",
-                                 package = "statgenSTA"),
-                     stringsAsFactors = FALSE, na.strings = c("NA", "*"))
+SB_Yield <- read.csv("./data-raw/SB_yield.csv", stringsAsFactors = FALSE,
+                     na.strings = c("NA", "*"))
 # Restrict to HEAT05
 Heat05 <- SB_Yield[SB_Yield[["Env"]] == "HEAT05", ]
 # Create object of class TD
@@ -17,9 +16,12 @@ usethis::use_data(TDHeat05, overwrite = TRUE)
 ## Create TDMaize.
 
 # Read raw data
-F2Maize <- read.csv(system.file("extdata", "F2maize_pheno.csv",
-                                package = "statgenSTA"),
-                    stringsAsFactors = FALSE)
+F2Maize <- read.csv("./data-raw/F2maize_pheno.csv", stringsAsFactors = FALSE)
+# Add column regime
+F2Maize[["regime"]] <- factor(substr(F2Maize[["env."]], 1, 2),
+                              levels = c("NS", "IS", "SS", "LN", "HN"))
+# Reorder columns.
+F2Maize <- F2Maize[, c("env.", "genotype.", "regime", "yld")]
 # Create object of class TD
 TDMaize <- createTD(data = F2Maize, genotype = "genotype.", trial = "env.")
 # Export to package
@@ -62,45 +64,4 @@ extractOptions <- read.csv("data-raw/extractOptions.csv",
 ## Export all internal data in one go to package.
 usethis::use_data(testData, testTD, extractOptions,
                   overwrite = TRUE, internal = TRUE)
-
-## Create data for vignette.
-# Read raw data.6
-dat2011 <- read.delim(system.file("extdata", "pheno_data2011.txt",
-                                  package = "statgenSTA"))
-dat2012 <- read.delim(system.file("extdata", "pheno_data2012.txt",
-                                  package = "statgenSTA"))
-# Split data into separate year/trials.
-dat2011_1 <- dat2011[, 1:11]
-dat2011_1[["trial"]] <- "SR_FI_11"
-colnames(dat2011_1)[8:11] <- c("DH", "GY", "NKS", "TKW")
-dat2011_2 <- dat2011[, c(1:7, 12:15)]
-dat2011_2[["trial"]] <- "SR_MWS_11"
-colnames(dat2011_2)[8:11] <- c("DH", "GY", "NKS", "TKW")
-dat2011tot <- rbind(dat2011_1, dat2011_2)
-dat2011tot[["year"]] <- 2011
-
-dat2012_1 <- dat2012[, 1:8]
-dat2012_1[["trial"]] <- "SR_FI_12"
-colnames(dat2012_1)[8] <- "GY"
-dat2012_2 <- dat2012[, c(1:7, 9)]
-dat2012_2[["trial"]] <- "SR_MWS_12"
-colnames(dat2012_2)[8] <- "GY"
-dat2012_3 <- dat2012[, c(1:7, 10)]
-dat2012_3[["trial"]] <- "C_SWS_12"
-colnames(dat2012_3)[8] <- "GY"
-dat2012tot <- rbind(dat2012_1, dat2012_2, dat2012_3)
-dat2012tot[["year"]] <- 2012
-dat2012tot[c("DH", "NKS", "TKW")] <- NA
-# Bind year data together and rename genotypes.
-wheatChl <- rbind(dat2011tot, dat2012tot)
-wheatChl$trt <- sprintf("G%03d", wheatChl[["trt_id"]])
-wheatChl <- wheatChl[!colnames(wheatChl) %in% c("parc", "trt_id")]
-# Export to package
-usethis::use_data(wheatChl, overwrite = TRUE)
-
-
-
-
-
-